Create basic files for page, simple readme intro
Because there was nothing here, I created a simple index.html file with
a boilerplate and a script-section, I created a styles.css which I
linked in the html-file. That's all.

The JavaScript goes entirely into the index.html and the styles.css I
need to create a basic styling of the page.