Create a reset button and a pixel resolution button
Because the user needs some interactivity, I'm sure he wants to
comfortably erase his drawing if he wants to draw something new.

So there is an erase button that resets everything and there is a
resolution button that prompts the user to enter a number between 1 - 100
that will hopefully change the number of squares in the grid and then
reloads the whole grid.