Flexbox for buttons and createGrid() function
Because the buttons were a mess, I scripted some flexbox formatting
for the buttons and also the grid to position everything nice and tidy.

Also I realized, that the grid-creation has to go into a function, that
the function can be called every time the user prompts the number of
rows and columns.