Make resolution button work properly
The resolution button now works as intended. The user is prompted to
enter a number between 1 and 100 that creates now a grid according to
input and also the grid is activated for the mouseover event listener.

Unfortunately, with these changes, the "Reset" button event listener
and the "Space Key" event listener, both resetting the grid to white, don't
work anymore. I have to find out, why.