Make the Reset button work as intended
The "Reset" button now actually resets the grid and the same can be said
when pressing the "Space Key" which also resets the grid.

But there's a bug. My first official bug: Pressing the "Space Key"
sometimes calls the "click" event listener for the Resolution button.

May have something to do with the "Space Key" also being an input to
press buttons as a default browser function.